requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
google-cloud-bigquery==3.23.0
google-auth==2.29.0
pandas==2.2.1
plotly==5.21.0
numpy==1.26.4
gspread==5.12.4
sqlalchemy==2.0.28
openai==1.16.0
python-dotenv==1.0.1
langchain==0.1.11
requests==2.31.0
pymongo==4.6.1
jinja2==3.1.3
aiosqlite==0.19.0
httpx==0.27.0
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4 